vercel/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.6
django-cors-headers==3.13.0
django-filter==23.3
django-jazzmin==2.6.0
django-js-asset==2.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
Pillow==10.0.1
#psycopg2==2.9.9
psycopg2-binary==2.9.5
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.6
gunicorn==21.1.0